Create SON index.html
Special Olympics is a global organization that serves athletes with intellectual disabilities working with hundreds of thousands of volunteers and coaches each year. Since the establishment of Special Olympics in 1968, the number of people with and without intellectual disabilities who are involved with the organization has been growing, but the unmet need to reach more people with intellectual disabilities is staggering.

It’s at the local level—right here—where interested volunteers meet the athletes. That’s where the perceptions start to change and where the miracle of transformation takes place.

Special Olympics provides year-round sports training and athletic competition in a variety of Olympic-type sports for children and adults with intellectual disabilities. Those activities give them continuing opportunities to develop physical fitness, demonstrate courage, experience joy and participate in a sharing of gifts, skills and friendship.